+++ gwcelery/tests/test_tasks_detchar.py
@@ -1,5 +1,5 @@
 import logging
-from unittest.mock import patch
+from unittest.mock import call, patch
 
 from pkg_resources import resource_filename
 import pytest
@@ -57,6 +57,39 @@ def test_check_idq(llhoft_glob_pass):
         'H1:IDQ-PGLITCH_OVL_32_2048', 0)
 
 
+@patch('time.strftime', return_value='00:00:00 UTC Mon 01 Jan 2000')
+@patch('socket.gethostname', return_value='test_host')
+@patch('getpass.getuser', return_value='test_user')
+def test_dqr_json(mock_time, mock_host, mock_user):
+    state = "pass"
+    summary = "72 and sunny!!"
+    assert detchar.dqr_json(state, summary) == {
+        'state': 'pass',
+        'process_name': 'gwcelery.tasks.detchar',
+        'process_version': 'unknown',
+        'librarian': 'Geoffrey Mo (geoffrey.mo@ligo.org)',
+        'date': '00:00:00 UTC Mon 01 Jan 2000',
+        'hostname': 'test_host',
+        'username': 'test_user',
+        'summary': '72 and sunny!!',
+        'figures': [],
+        'tables': [],
+        'links': ([
+            {
+                "href":
+                "https://gwcelery.readthedocs.io/en/latest/gwcelery.tasks.detchar.html#gwcelery.tasks.detchar.check_vectors", # noqa
+                "innerHTML": "a link to the documentation for this process"
+            },
+            {
+                "href":
+                "https://git.ligo.org/emfollow/gwcelery/blob/master/gwcelery/tasks/detchar.py",  # noqa
+                "innerHTML": "a link to the source code in the gwcelery repo"
+            }
+        ],),
+        'extra': []
+    }
+
+
 def test_check_vector(llhoft_glob_pass):
     channel = 'H1:DMT-DQ_VECTOR'
     start, end = 1216577976, 1216577980
@@ -75,19 +108,41 @@ def test_check_vectors_skips_mdc(caplog):
     assert 'Skipping state vector checks because M1234 is an MDC' in messages
 
 
+@patch('gwcelery.tasks.detchar.dqr_json', return_value='dqrjson')
 @patch('gwcelery.tasks.gracedb.client.writeLog')
 @patch('gwcelery.tasks.gracedb.create_label')
-def test_check_vectors(mock_create_label, mock_write_log,
+def test_check_vectors(mock_create_label, mock_write_log, mock_json,
                        llhoft_glob_pass, ifo_h1, ifo_h1_idq):
     event = {'search': 'AllSky', 'instruments': 'H1', 'pipeline': 'oLIB'}
     superevent_id = 'S12345a'
     start, end = 1216577977, 1216577979
     detchar.check_vectors(event, superevent_id, start, end)
-    mock_write_log.assert_called_with(
-        'S12345a',
-        ('detector state for active instruments is good. For all instruments,'
-         ' bits good (H1:NO_OMC_DCPD_ADC_OVERFLOW,'
-         ' H1:NO_DMT-ETMY_ESD_DAC_OVERFLOW, H1:HOFT_OK,'
-         ' H1:OBSERVATION_INTENT), bad (), unknown().'),
-        tag_name=['data_quality'])
+    print(mock_write_log.mock_calls)
+    calls = [
+        call(
+            'S12345a',
+            ('detector state for active instruments is good.'
+             ' For all instruments,'
+             ' bits good (H1:NO_OMC_DCPD_ADC_OVERFLOW,'
+             ' H1:NO_DMT-ETMY_ESD_DAC_OVERFLOW, H1:HOFT_OK,'
+             ' H1:OBSERVATION_INTENT), bad (), unknown().'),
+            tag_name=['data_quality']),
+        call(
+            'S12345a',
+            'No HW injections found.',
+            tag_name=['data_quality']),
+        call(
+            'S12345a',
+            ('iDQ glitch probabilities at both H1 and L1'
+             ' are good (below {}).').format(
+                 app.conf['idq_pglitch_thresh']),
+            tag_name=['data_quality']),
+        call(
+            'S12345a',
+            'DQR-compatible json generated from check_vectors results',
+            'gwcelerydetcharcheckvectors-S12345a.json',
+            '"dqrjson"',
+            ['data_quality']),
+    ]
+    mock_write_log.assert_has_calls(calls, any_order=True)
     mock_create_label.assert_called_with('DQOK', 'S12345a')